feat: zero query being used now

+++ packages/zero/src/scripts/clean.ts
@@ -2,16 +2,19 @@ import { exec } from "../index";
 
 console.log("Cleaning up resources...");
 
-try {
-  exec("rm -f .tmp/next-oral.db*");
-} catch (err) {
-  console.info((err as Error).message);
-}
+// eslint-disable-next-line turbo/no-undeclared-env-vars
+if (process.env.CLEAN_ZERO) {
+  try {
+    exec("rm -f .tmp/next-oral.db*");
+  } catch (err) {
+    console.info((err as Error).message);
+  }
 
-try {
-  exec("docker rm -f next-oral");
-} catch (err) {
-  console.info((err as Error).message);
+  try {
+    exec("docker rm -f next-oral");
+  } catch (err) {
+    console.info((err as Error).message);
+  }
 }
 
 console.log("Cleanup complete.");
